// src/types/form.types.ts
export type FieldType = "text" | "dropdown" | "radio" | "checkbox";

export type TextFieldType =
  | "text"
  | "email"
  | "phone"
  | "phone_with_code"
  | "password"
  | "number"
  | "decimal"
  | "otp"
  | "url"
  | "search"
  | "multiline"
  | "masked"
  | "custom";

export interface CountryCodeOption {
  code: string;
  label: string;
  dialCode: string;
}

export interface CountryCodePickerStyle {
  backdrop?: ViewStyle;
  modal?: ViewStyle;
  line?: ViewStyle;
  searchMessageText?: TextStyle;
  itemsList?: ViewStyle;
  modalInner?: ViewStyle;
  countryMessageContainer?: ViewStyle;
  textInput?: TextStyle;
  countryButtonStyles?: ViewStyle;
  flag?: TextStyle;
  dialCode?: TextStyle;
  countryName?: TextStyle;
}

export interface BaseFieldSchema {
  name: string;
  label: string;
  type: FieldType;
  required?: boolean;
  requiredMessage?: string;
}

export interface TextFieldSchema extends BaseFieldSchema {
  type: "text";
  textFieldType?: TextFieldType;
  placeholder?: string;
  emailRegex?: RegExp;
  emailValidationMessage?: string;
  validationRegex?: RegExp;
  validationMessage?: string;
  textInputProps?: TextInputProps;
  containerStyle?: StyleProp<ViewStyle>;
  phoneWithCodeContainerStyle?: StyleProp<ViewStyle>;
  countryCodeButtonStyle?: StyleProp<ViewStyle>;
  countryCodeTextStyle?: StyleProp<TextStyle>;
  countryCodeDropdownStyle?: StyleProp<ViewStyle>;
  countryCodeItemStyle?: StyleProp<ViewStyle>;
  countryCodeItemTextStyle?: StyleProp<TextStyle>;
  countryCodePickerStyle?: CountryCodePickerStyle;
  labelStyle?: StyleProp<TextStyle>;
  inputStyle?: StyleProp<TextStyle>;
  errorTextStyle?: StyleProp<TextStyle>;
  countryCodes?: CountryCodeOption[];
  defaultCountryCode?: string;
  countryCodeFieldName?: string;
  passwordVisibilityButtonStyle?: StyleProp<ViewStyle>;
  passwordVisibilityTextStyle?: StyleProp<TextStyle>;
  renderEyeIcon?: (visible: boolean) => React.ReactNode;
}
// ...
export type FormValues = Record<string, any>;